requirements.txt
git+https://github.com/almottier/TapoP100.git@main
PyYAML~=6.0.1
python-dotenv~=1.0.1
discord-webhook~=1.3.1